+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # next
 
-- Feature: Add subscribable syncStatuses map to SyncCoordinator, with number of
-  ingested docs and 'caught up' status of each syncing session.
+- Feature: Syncer.syncStatuses. Subscribable map of syncer's connections' sync
+  statuses.
+- Feature: SyncCoordinator.syncStatuses. Subscribable map of coordinator's
+  shares' sync statuses, with number of ingested docs and 'caught up' status of
+  each syncing session.
 - Patch: Common shares between peers are re-established whenever a Peer's set of
   replicas chages.